refactor(Authentication): BasicAuthInterceptor app 패키지 의존성 제거 및 분리
